L12 compatibility (#81)
This pull request includes updates to the workflow configuration and
dependency management to support newer versions of PHP and Laravel, as
well as improvements to the test execution process.

### Workflow Configuration Updates:

*
[`.github/workflows/run-tests.yml`](diffhunk://#diff-7314d0ebbd2e9537ae4889316745b4fd2fa43cb86275c9caae18a86ba228b642L7-R47):
Updated the PHP and Laravel versions matrix to include PHP 8.4 and
Laravel ^12.0. The matrix now also includes an option to run tests with
or without coverage.
*
[`.github/workflows/run-tests.yml`](diffhunk://#diff-7314d0ebbd2e9537ae4889316745b4fd2fa43cb86275c9caae18a86ba228b642R67-R77):
Modified the test execution steps to conditionally run tests with or
without coverage based on the matrix configuration.

### Dependency Management Updates:

*
[`composer.json`](diffhunk://#diff-d2ab9925cad7eac58e0ff4cc0d251a937ecf49e4b6bf57f8b95aab76648a9d34L18-R23):
Added support for Laravel ^12.0 and Pest ^3.7 in the `require` and
`require-dev` sections, respectively.
*
[`composer.json`](diffhunk://#diff-d2ab9925cad7eac58e0ff4cc0d251a937ecf49e4b6bf57f8b95aab76648a9d34L53-R55):
Set the minimum stability to "dev" and preferred stability to true to
ensure the use of stable versions when available.

---------

Co-authored-by: Shift <shift@laravelshift.com>